Split unit tests to have them run without infra required
