+++ .github/workflows/ci-test-api.yml
@@ -66,7 +66,7 @@ jobs:
             cp /home/workspace/opencti-platform/.yarnrc.yml .
             yarn install
             yarn build
-            NODE_OPTIONS=--max_old_space_size=8192 yarn test:ci-integration
+            NODE_OPTIONS=--max_old_space_size=8192 yarn test:ci-integration  --coverage
         '
 
     - name: Upload backend test result
@@ -167,7 +167,7 @@ jobs:
             cp /home/workspace/opencti-platform/.yarnrc.yml .
             yarn install
             yarn build
-            NODE_OPTIONS=--max_old_space_size=8192 yarn test:ci-rules
+            NODE_OPTIONS=--max_old_space_size=8192 yarn test:ci-rules-and-others --coverage
         '
 
     # Output all logs used during the job
@@ -222,7 +222,6 @@ jobs:
           sh -c 'set -ex;
             echo "$(ls /root/.yarn/berry/cache/ 2> /dev/null | wc -l || echo 0) packages in Yarn global cache";
             apk add build-base git libffi-dev cargo;
-            git clone https://github.com/OpenCTI-Platform/connectors.git /home/workspace/opencti-connectors  # TODO Remove when not anymore needed to run integration test
             cd /home/workspace/opencti-platform/opencti-graphql
             npm install -g corepack
             cp /home/workspace/opencti-platform/.yarnrc.yml .
@@ -230,7 +229,7 @@ jobs:
             yarn build
             yarn check-ts
             yarn lint
-            NODE_OPTIONS=--max_old_space_size=8192 yarn test:ci-unit-and-others
+            NODE_OPTIONS=--max_old_space_size=8192 yarn test:ci-unit --coverage
         '
 
     # Output all logs used during the job
