+++ opencti-platform/opencti-graphql/package.json
@@ -26,12 +26,12 @@
     "test:dev:init": "INIT_TEST_PLATFORM=true vitest run --config vitest.config.dev.ts --testNamePattern=\"Should import creation succeed\" tests/02-integration/00-inject/loader-test.ts",
     "test:dev:resume": "SKIP_CLEANUP_PLATFORM=true vitest run --config vitest.config.dev.ts",
     "test:dev": "vitest run --config vitest.config.dev.ts",
-    "test:ci-unit-and-others": "vitest run --config vitest.config.ci-unit-and-others.ts",
-    "test:ci-integration": "vitest run --config vitest.config.ci-integration.ts",
-    "test:ci-rules": "vitest run --config vitest.config.ci-rules.ts",
+    "test:ci-unit": "vitest --bail 1 run --config vitest.config.ci-unit.ts",
+    "test:ci-integration": "vitest --bail 1 run --config vitest.config.ci-integration.ts",
+    "test:ci-rules-and-others": "vitest --bail 1 run --config vitest.config.ci-rules-and-others.ts",
     "test:watch:resume": "SKIP_CLEANUP_PLATFORM=true vitest watch --config vitest.config.dev.ts",
     "test:watch": "vitest watch --config vitest.config.dev.ts",
-    "test": "vitest --silent --bail 1 run --config vitest.config.test.ts --coverage",
+    "test": "vitest --silent --bail 1 run --config vitest.config.integration.ts --coverage",
     "verify-licenses": "npx license-checker-rseidelsohn --summary --onlyAllow 'MIT;Apache-2.0;GPL-2.0-or-later;OFL-1.1;BSD-3-Clause;ISC;BSD-2-Clause;Python-2.0;MPL-2.0;CC-BY-4.0;CC-BY-3.0;BlueOak-1.0.0;CC0-1.0;UNLICENSED;0BSD;Unlicense;WTFPL' --excludePackagesStartingWith  'eslint;opencti'",
     "wait-api": "node build/script-wait-for-api.js",
     "get-connectors-manifest": "node script/get-connectors-manifest.js"
