+++ opencti-platform/opencti-graphql/tests/02-integration/01-database/confidence-level-test.ts
