+++ opencti-platform/opencti-graphql/tests/02-integration/01-database/schema-attributes-test.ts
