+++ opencti-platform/opencti-graphql/tests/02-integration/01-database/xtm-hub-test.ts
