+++ opencti-platform/opencti-graphql/tests/02-integration/04-manager/fileIndexManager-test.ts
