+++ opencti-platform/opencti-graphql/vitest.config.ci-integration.ts
@@ -1,4 +1,4 @@
-/* eslint-disable import/no-extraneous-dependencies */
-import { buildTestConfig } from './vitest.config.test';
+ 
+import { buildIntegrationTestConfig } from './vitest.config.integration';
 
-export default buildTestConfig(['tests/(02|03|04)-*/**/*-test.{js,mjs,cjs,ts,mts,cts,jsx,tsx}']);
+export default buildIntegrationTestConfig(['tests/(02|03|04)-*/**/*-test.{js,mjs,cjs,ts,mts,cts,jsx,tsx}']);
