+++ opencti-platform/opencti-graphql/vitest.config.ci-unit.ts
@@ -1,0 +1,4 @@
+ 
+import { buildTestConfig } from './vitest.config';
+
+export default buildTestConfig(['tests/(01)-*/**/*-test.{js,mjs,cjs,ts,mts,cts,jsx,tsx}']);
